feat: version form and reorganize common directory by epic
