+++ client/src/common/epic/my-passport-form/types/MyPassportFormVersion.enum.ts
@@ -1,0 +1,3 @@
+export enum MyPassportFormVersion {
+	V1_0 = "1.0",
+}
