+++ client/src/common/epic/my-passport-form/types/form.ts
@@ -1,6 +1,9 @@
+import type { MyPassportFormVersion } from "./MyPassportFormVersion.enum";
+
 export type MyPassportForm = {
 	personalDetails: PersonalDetails;
 	addressDetails: AddressDetails;
+	version?: MyPassportFormVersion;
 };
 
 export type PersonalDetails = {
