+++ client/src/feat/my-passport-form/types/index.ts
@@ -1,4 +1,4 @@
-export * from "common/types/epic/my-passport-form";
+export * from "common/epic/my-passport-form/types";
 export * from "./api";
 export * from "./country.enum";
 export * from "./external";
