+++ client/src/feat/my-passport-form/types/ui.ts
@@ -1,5 +1,5 @@
 import type { createForm, FormStore } from "@modular-forms/solid";
-import type { MyPassportForm } from "common/types/epic/my-passport-form";
+import type { MyPassportForm } from "common/epic/my-passport-form/types";
 import type { Accessor } from "solid-js";
 import type { DropdownOptions } from "./api";
